+++ Firestore/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Unreleased
+- [changed] Replaced the `DocumentListenOptions` object with a simple boolean.
+  Instead of calling
+  `addSnapshotListener(options: DocumentListenOptions.includeMetadataChanges(true))`
+  call `addSnapshotListener(includeMetadataChanges:true)`.
 - [changed] Replaced the `SetOptions` object with a simple boolean. Instead of
   calling `setData(["a": "b"], options: SetOptions.merge())` call
   `setData(["a": "b"], merge: true)`.